requirements.txt
tensorflow>=2.13.0
tf-keras>=2.16.0
opencv-python>=4.8.0
numpy<2
Pillow>=10.0.0
scikit-learn>=1.3.0
streamlit>=1.28.0
ollama>=0.4.0
chromadb>=0.4.0
